refactor: change @types to @models
