+++ src/pages/index.tsx
@@ -1,4 +1,4 @@
-import { File, MdxEdge } from '@types';
+import { File, MdxEdge } from '@models';
 import { graphql, PageProps } from 'gatsby';
 import React from 'react';
 import { Layout } from '../components/common/layout';
