+++ src/providers/color-mode-provider.tsx
@@ -1,4 +1,4 @@
-import { ColorMode } from '@types';
+import { ColorMode } from '@models';
 import React from 'react';
 
 export type ColorModeContextValue = {
@@ -8,7 +8,7 @@ export type ColorModeContextValue = {
 
 const defaultContext: ColorModeContextValue = {
   colorMode: 'light',
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
+  // eslint-disable-next-line @modelscript-eslint/no-empty-function
   setColorMode: () => {},
 };
